Image_classification_Dog_cat_Celebal-project/AnimalPlant_Classifier/requirements.txt
streamlit
joblib
numpy
pandas
matplotlib
scikit-learn==1.6.1
Pillow
xgboost
